[REFACTOR] Add PageService to move checkIfPageShouldBeSkipped out of the CrawlerController
