+++ phpstan-baseline.neon
@@ -313,7 +313,7 @@ parameters:
 			path: Classes/Controller/CrawlerController.php
 
 		-
-			message: "#^Class cognitive complexity is 279, keep it under 60$#"
+			message: "#^Class cognitive complexity is 267, keep it under 60$#"
 			count: 1
 			path: Classes/Controller/CrawlerController.php
